requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
lxml==4.6.3
scraperapi-sdk==0.2.2
pymongo==4.1.1
aiohttp
asyncio